+++ src/ApiResources/AnonymousResourceTypeCollection.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace CbtechLtd\JsonApiTransformer\ApiResources;
+
+class AnonymousResourceTypeCollection extends ResourceTypeCollection
+{
+    public function __construct(array $items = [])
+    {
+        $this->items = $items;
+    }
+
+    public function resourceType(): string
+    {
+        return '';
+    }
+}