+++ src/ApiResources/ApiResourceCollection.php
@@ -2,6 +2,8 @@
 
 namespace CbtechLtd\JsonApiTransformer\ApiResources;
 
+use CbtechLtd\JsonApiTransformer\Contracts\ResourceLinkContract;
+use CbtechLtd\JsonApiTransformer\Contracts\ResourceMetaContract;
 use Illuminate\Http\Resources\Json\ResourceCollection;
 use Illuminate\Support\Collection;
 use Webmozart\Assert\Assert;
@@ -9,31 +11,53 @@ use Webmozart\Assert\Assert;
 class ApiResourceCollection extends ResourceCollection
 {
     public $collects = ApiResource::class;
+    protected array $links = [];
+    protected array $meta = [];
 
     public static function make(...$parameters)
     {
-        return static::makeWithType($parameters[0], $parameters[1]);
+        return new static($parameters[0]);
     }
 
-    public static function makeWithType(string $resourceType, Collection $items)
+    public static function makeFromModels(string $resourceType, Collection $items)
     {
         Assert::isAOf($resourceType, ResourceType::class, 'First parameter must be a ' . ResourceType::class);
 
-        return new static(
+        return static::make(
             $items->map(fn($it) => new $resourceType($it))
         );
+    }
+
+    public function withMeta(array $meta): self
+    {
+        Assert::allIsInstanceOf(
+            $meta,
+            ResourceMetaContract::class,
+            'All items must be an instance of ' . ResourceMetaContract::class
+        );
+
+        $this->meta = $meta;
+        return $this;
+    }
+
+    public function withLinks(array $links): self
+    {
+        Assert::allIsInstanceOf(
+            $links,
+            ResourceLinkContract::class,
+            'All items must be an instance of ' . ResourceLinkContract::class
+        );
+
+        $this->links = $links;
+        return $this;
     }
 
     public function toArray($request)
     {
         return [
             'data'  => $this->getData($request),
-
-            // TODO: Build the links object.
-            'links' => [],
-
-            // TODO: Build the meta object.
-            'meta'  => [],
+            'links' => $this->buildLinks(),
+            'meta'  => $this->buildMeta(),
         ];
     }
 
@@ -46,6 +70,20 @@ class ApiResourceCollection extends ResourceCollection
     {
         return $this->collection->map(
             fn(ApiResource $resource) => $resource->toArray($request)
+        );
+    }
+
+    protected function buildLinks(): Collection
+    {
+        return Collection::make($this->links)->mapWithKeys(
+            fn(ResourceLinkContract $link) => $link->toArray()
+        );
+    }
+
+    protected function buildMeta(): Collection
+    {
+        return Collection::make($this->meta)->mapWithKeys(
+            fn(ResourceMetaContract $meta) => $meta->toArray()
         );
     }
 }