+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to `laravel-json-api-transformer` will be documented in this file
 
-## 1.0.0 - 2020-05-29
+## 0.2.0 - 2020-08-11
+
+### Added
+- Links and meta processing in resource collections.
+- Resource entry collection.
+- API transformer method to build from models.
+
+### Changed
+- API Transformer collection method.
+
+---
+
+## 0.1.0 - 2020-05-29
 
 - initial release